Fixed bug in constrained exp_imp for search and infill, tested and working now

+++ src/case_spec.py
@@ -50,13 +50,14 @@ bu_steps = (0.0, 5.0, 89.0, 183.0)
 # '~joshrich', 'SERPENT', 'new_core', 'opt_runs_new'
 #data_dir = os.path.join('~joshrich', 'SERPENT', 'new_core', 'opt_runs_rand')
 data_dir = os.path.join('~joshrich', 'SERPENT', 'new_core', 'opt_runs_pow')
-dump_dir = os.path.join(data_dir, 'run_dump_files', 'lhs_10_test1')
-
+# 'run_dump_files', 'lhs_10_test1'
+# 'test_exec', 'test_efpd_neg', 'lhs_110_efpd_testdata'
+dump_dir = os.path.join(data_dir, 'run_dump_files', 'lhs_110_test1')
 
 run_opts = dict([('fuel_xs', '.15c'), ('mod_xs','.12c'),('cool_xs','.09c'), ('pin_rad','0.7'), \
                  ('cool_mat', 'flibe'), ('sab_xs', '.24t'),('mod_sab_xs', '.22t'), ('total_coreh','175')])
                  
-doe_opts = {'doe_type':'LHS', 'num_LHS_samples':10, 'LHS_type':'maximin'}  # {'doe_type':'FF', 'FF_num':3}, {'doe_type':'LHS', 'num_LHS_samples':20, 'LHS_type':'maximin'}
+doe_opts = {'doe_type':'O-LHS', 'num_LHS_samples':110, 'LHS_type':'maximin'}  # {'doe_type':'FF', 'FF_num':3}, {'doe_type':'LHS', 'num_LHS_samples':20, 'LHS_type':'maximin'}
 
                  
 doe_sets = {}
@@ -114,9 +115,9 @@ converge_opts = {'converge_tol':1e-3, 'converge_points':3,
                  'converge_type':'rel'}
 thresh_in = 1e-3
 euclid_tol = 1e-3
-outp_mode = 'interact' # either 'interact' or 'iterate'
-run_mode = 'normal' # either 'restart' or 'normal'
-use_exist_data = 'off'
+outp_mode = 'iterate' # either 'interact' or 'iterate'
+run_mode = 'restart' # either 'restart' or 'normal'
+use_exist_data = 'on'
 submit_interval = 10
 
 
@@ -125,6 +126,8 @@ timenow = time.localtime()
 timestring = '_{}_{}_{}_{}_{}_{}'.format(*timenow[0:6])
 
 if run_mode == 'normal':
+    if use_exist_data == 'on':
+        raise Exception("If run_mode = 'normal', use_exist_data cannot be 'on' ")
     try:
         namestring = data_opts['data_fname'][:-4] + timestring + '.out'
         os.rename(data_opts['data_fname'], namestring)
